Fix dependency conflicts, migrate Riverpod 3.x, and fix Phase 1&2 bugs
- Upgrade 28 dependencies to latest major versions (Firebase 4.x, Riverpod 3.x,
  go_router 17.x, geolocator 14.x, freezed 3.x, etc.)
- Migrate all 13 providers from removed StateNotifier to Notifier pattern
- Fix phone validation: India-specific +91 support, E.164 normalization,
  unicode name support with hyphens/apostrophes
- Fix auth flow: add resendOtp(), persist phone number in state, map Firebase
  errors to user-friendly messages, add phone validator to input
- Fix OTP screen: clear fields on error, implement actual resend logic
- Fix null-safe Firestore deserialization across all models (Timestamp, String)
- Fix router: auth stream loading guard, kDebugMode for debug logs,
  validate OTP verificationId exists before rendering
- Persist onboarding/profile completion state to Hive local storage
- Gate all logger levels behind kDebugMode, add PII auto-redaction
- Remove requestCamera() (violates no-camera architecture rule)
- Fix notification ID collision with incrementing counter
- Fix battery service race condition on _hasAlerted flag
- Fix shake service duplicate subscription prevention
- Fix hashCode null safety in Failure base class
- Fix string_extensions initials crash on multi-space strings
- Clamp threat score to 0-100 range

+++ lib/core/services/location_service.dart
@@ -44,7 +44,7 @@ class LocationService {
 
   void startTracking({
     int intervalMs = 10000,
-    double distanceFilter = 5,
+    int distanceFilter = 5,
   }) {
     _positionSubscription?.cancel();
     _positionSubscription = Geolocator.getPositionStream(
